fix: int to long in lfhcal_studiesProcessor.cc

+++ src/benchmarks/reconstruction/lfhcal_studies/lfhcal_studiesProcessor.cc
@@ -329,7 +329,7 @@ void lfhcal_studiesProcessor::Process(const std::shared_ptr<const JEvent>& event
     auto detector_passive = m_decoder->get(cellID, iPassive);
     auto detector_layer_x = m_decoder->get(cellID, iLx);
     auto detector_layer_y = m_decoder->get(cellID, iLy);
-    int detector_layer_rz = -1;
+    long detector_layer_rz = -1;
     if (isLFHCal)
       detector_layer_rz = m_decoder->get(cellID, 7);
     auto detector_layer_z = m_decoder->get(cellID, iLz);
@@ -341,9 +341,9 @@ void lfhcal_studiesProcessor::Process(const std::shared_ptr<const JEvent>& event
 
     if (detector_passive > 0) continue;
     // calc cell IDs
-    int cellIDx = -1;
-    int cellIDy = -1;
-    int cellIDz = -1;
+    long cellIDx = -1;
+    long cellIDy = -1;
+    long cellIDz = -1;
     if (isLFHCal){
       cellIDx = 54*2 - detector_module_x * 2 + detector_layer_x;
       cellIDy = 54*2 - detector_module_y * 2 + detector_layer_y;
